remove all box<dyn>
