+++ Cargo.toml
@@ -10,7 +10,7 @@ publish = false
 crate-type = ["cdylib"]
 
 [features]
-default = ["build3d"]
+default = ["build2d"]
 build2d = ["single-dim2", "serde-serialize", "parallel", "simd-stable"]
 build3d = ["single-dim3", "serde-serialize", "parallel", "simd-stable"]
 dim2 = []
