+++ src/bodies/mod.rs
@@ -1,6 +1,7 @@
 pub mod rapier_area;
 pub mod rapier_body;
 pub mod rapier_collision_object;
+pub mod rapier_collision_object_base;
 pub mod rapier_collision_object_impl;
 #[cfg(feature = "dim2")]
 pub mod rapier_direct_body_state_2d;
