+++ src/joints/mod.rs
@@ -2,11 +2,13 @@
 pub mod rapier_cone_twist_joint_3d;
 #[cfg(feature = "dim2")]
 pub mod rapier_damped_spring_joint_2d;
+pub mod rapier_empty_joint;
 #[cfg(feature = "dim3")]
 pub mod rapier_generic_6dof_joint_3d;
 #[cfg(feature = "dim2")]
 pub mod rapier_groove_joint_2d;
 pub mod rapier_joint;
+pub mod rapier_joint_base;
 pub mod rapier_revolute_joint;
 #[cfg(feature = "dim3")]
 pub mod rapier_slider_joint_3d;
