+++ src/joints/rapier_slider_joint_3d.rs
@@ -1,9 +1,10 @@
 use godot::classes::*;
 use godot::prelude::*;
 
-use crate::bodies::rapier_collision_object::IRapierCollisionObjectBase;
+use super::rapier_joint_base::RapierJointBase;
+use crate::bodies::rapier_collision_object::IRapierCollisionObject;
+use crate::bodies::rapier_collision_object::RapierCollisionObject;
 use crate::joints::rapier_joint::IRapierJoint;
-use crate::joints::rapier_joint::RapierJointBase;
 use crate::rapier_wrapper::prelude::*;
 #[cfg_attr(
     feature = "serde-serialize",
@@ -18,8 +19,8 @@ impl RapierSliderJoint3D {
     pub fn new(
         anchor_a: Transform3D,
         anchor_b: Transform3D,
-        body_a: &Box<dyn IRapierCollisionObjectBase>,
-        body_b: &Box<dyn IRapierCollisionObjectBase>,
+        body_a: &RapierCollisionObject,
+        body_b: &RapierCollisionObject,
         physics_engine: &mut PhysicsEngine,
     ) -> Self {
         let invalid_joint = Self {
