+++ src/servers/rapier_physics_server_extra.rs
@@ -3,6 +3,7 @@ use godot::global::rid_from_int64;
 use godot::prelude::*;
 
 use super::rapier_physics_singleton::physics_data;
+use crate::bodies::rapier_collision_object::IRapierCollisionObject;
 use crate::fluids::rapier_fluid::RapierFluid;
 use crate::servers::RapierPhysicsServer;
 use crate::types::*;
