+++ src/servers/rapier_physics_singleton.rs
@@ -1,16 +1,16 @@
 use godot::prelude::*;
 use hashbrown::HashMap;
 
-use crate::bodies::rapier_collision_object::{IRapierCollisionObjectBase, RapierCollisionObjectBase};
+use crate::bodies::rapier_collision_object::RapierCollisionObject;
 use crate::fluids::rapier_fluid::RapierFluid;
 use crate::joints::rapier_joint::RapierJoint;
 use crate::rapier_wrapper::prelude::*;
-use crate::shapes::rapier_shape::{IRapierShape, RapierShape};
+use crate::shapes::rapier_shape::RapierShape;
 use crate::spaces::rapier_space::RapierSpace;
 pub type PhysicsShapes = HashMap<Rid, RapierShape>;
 pub type PhysicsSpaces = HashMap<Rid, RapierSpace>;
 pub type PhysicsActiveSpaces = HashMap<WorldHandle, Rid>;
-pub type PhysicsCollisionObjects = HashMap<Rid, RapierCollisionObjectBase>;
+pub type PhysicsCollisionObjects = HashMap<Rid, RapierCollisionObject>;
 pub type PhysicsJoints = HashMap<Rid, RapierJoint>;
 pub type PhysicsFluids = HashMap<Rid, RapierFluid>;
 #[derive(Default)]
