+++ src/shapes/rapier_concave_polygon_shape.rs
@@ -6,6 +6,7 @@ use godot::prelude::*;
 
 use crate::rapier_wrapper::prelude::*;
 use crate::shapes::rapier_shape::*;
+use crate::shapes::rapier_shape_base::RapierShapeBase;
 use crate::types::PackedVectorArray;
 #[cfg_attr(
     feature = "serde-serialize",
