+++ src/shapes/rapier_convex_polygon_shape.rs
@@ -6,7 +6,7 @@ use godot::prelude::*;
 
 use crate::rapier_wrapper::prelude::*;
 use crate::shapes::rapier_shape::IRapierShape;
-use crate::shapes::rapier_shape::RapierShapeBase;
+use crate::shapes::rapier_shape_base::RapierShapeBase;
 use crate::types::*;
 #[cfg_attr(
     feature = "serde-serialize",
