+++ src/shapes/rapier_cylinder_shape_3d.rs
@@ -3,6 +3,7 @@ use godot::prelude::*;
 
 use crate::rapier_wrapper::prelude::*;
 use crate::shapes::rapier_shape::*;
+use crate::shapes::rapier_shape_base::RapierShapeBase;
 #[cfg_attr(
     feature = "serde-serialize",
     derive(serde::Serialize, serde::Deserialize)
