+++ src/shapes/rapier_shape.rs
@@ -3,14 +3,22 @@ use godot::classes::physics_server_2d::*;
 #[cfg(feature = "dim3")]
 use godot::classes::physics_server_3d::*;
 use godot::prelude::*;
-use hashbrown::HashMap;
 
-use crate::rapier_wrapper::prelude::*;
-use crate::servers::rapier_physics_singleton::PhysicsData;
-use crate::types::*;
-
+use super::rapier_capsule_shape::RapierCapsuleShape;
+use super::rapier_circle_shape::RapierCircleShape;
+use super::rapier_concave_polygon_shape::RapierConcavePolygonShape;
+use super::rapier_convex_polygon_shape::RapierConvexPolygonShape;
+#[cfg(feature = "dim3")]
+use super::rapier_cylinder_shape_3d::RapierCylinderShape3D;
+#[cfg(feature = "dim3")]
+use super::rapier_heightmap_shape_3d::RapierHeightMapShape3D;
+use super::rapier_rectangle_shape::RapierRectangleShape;
+#[cfg(feature = "dim2")]
+use super::rapier_segment_shape_2d::RapierSegmentShape2D;
 use super::rapier_separation_ray_shape::RapierSeparationRayShape;
 use super::rapier_world_boundary_shape::RapierWorldBoundaryShape;
+use crate::rapier_wrapper::prelude::*;
+use crate::shapes::rapier_shape_base::RapierShapeBase;
 #[cfg_attr(feature = "serde-serialize", typetag::serde(tag = "type"))]
 pub trait IRapierShape {
     fn get_base(&self) -> &RapierShapeBase;
@@ -22,132 +30,101 @@ pub trait IRapierShape {
     fn get_data(&self) -> Variant;
     fn get_handle(&self) -> ShapeHandle;
 }
-pub enum RapierShape {
-    RapierSeparationRayShape(RapierSeparationRayShape),
-    RapierWorldBoundaryShape(RapierWorldBoundaryShape),
-}
 #[cfg_attr(
     feature = "serde-serialize",
     derive(serde::Serialize, serde::Deserialize)
 )]
-pub struct RapierShapeBase {
-    #[cfg_attr(feature = "serde-serialize", serde(skip, default = "invalid_rid"))]
-    rid: Rid,
-    aabb: Rect,
-    // TODO serialize this
-    #[cfg_attr(feature = "serde-serialize", serde(skip))]
-    owners: HashMap<Rid, i32>,
-    handle: ShapeHandle,
-}
-impl Default for RapierShapeBase {
-    fn default() -> Self {
-        Self {
-            rid: Rid::Invalid,
-            aabb: Rect::default(),
-            owners: HashMap::default(),
-            handle: ShapeHandle::default(),
-        }
-    }
-}
-impl RapierShapeBase {
-    pub(super) fn new(rid: Rid) -> Self {
-        Self {
-            rid,
-            aabb: Rect::default(),
-            owners: HashMap::default(),
-            handle: ShapeHandle::default(),
-        }
-    }
-
-    pub(super) fn set_handle(&mut self, handle: ShapeHandle, physics_engine: &mut PhysicsEngine) {
-        if self.handle != ShapeHandle::default() {
-            self.destroy_shape(physics_engine);
-        }
-        let rapier_aabb = physics_engine.shape_get_aabb(handle);
-        let vertices = rapier_aabb.vertices();
-        self.aabb = Rect::new(
-            vector_to_godot(vertices[0].coords),
-            vector_to_godot(rapier_aabb.extents()),
-        );
-        self.handle = handle;
-    }
-
-    pub fn get_handle(&self) -> ShapeHandle {
-        self.handle
-    }
-
-    pub fn is_valid(&self) -> bool {
-        self.handle != ShapeHandle::default()
-    }
-
-    pub fn call_shape_changed(
-        owners: HashMap<Rid, i32>,
-        shape_rid: Rid,
-        physics_data: &mut PhysicsData,
-    ) {
-        for (owner, _) in owners {
-            if let Some(owner) = physics_data.collision_objects.get_mut(&owner) {
-                owner.shape_changed(
-                    shape_rid,
-                    &mut physics_data.physics_engine,
-                    &mut physics_data.shapes,
-                    &mut physics_data.spaces,
-                );
-            }
-        }
-    }
-
-    pub fn get_aabb(&self, origin: Vector) -> Rect {
-        let mut aabb_clone = self.aabb;
-        aabb_clone.position += origin;
-        aabb_clone
-    }
-
-    #[cfg(feature = "dim2")]
-    pub fn get_aabb_area(&self) -> real {
-        self.aabb.area()
-    }
-
+pub enum RapierShape {
+    RapierCapsuleShape(RapierCapsuleShape),
+    RapierCircleShape(RapierCircleShape),
+    RapierConcavePolygonShape(RapierConcavePolygonShape),
+    RapierConvexPolygonShape(RapierConvexPolygonShape),
     #[cfg(feature = "dim3")]
-    pub fn get_aabb_area(&self) -> real {
-        self.aabb.volume()
-    }
+    RapierCylinderShape3D(RapierCylinderShape3D),
+    #[cfg(feature = "dim3")]
+    RapierHeightMapShape3D(RapierHeightMapShape3D),
+    RapierRectangleShape(RapierRectangleShape),
+    #[cfg(feature = "dim2")]
+    RapierSegmentShape2D(RapierSegmentShape2D),
+    RapierSeparationRayShape(RapierSeparationRayShape),
+    RapierWorldBoundaryShape(RapierWorldBoundaryShape),
+}
+macro_rules! impl_rapier_shape_trait {
+    ($enum_name:ident, $($variant:ident),*) => {
+        #[cfg_attr(feature = "serde-serialize", typetag::serde)]
+        impl IRapierShape for $enum_name {
+            fn get_base(&self) -> &RapierShapeBase {
+                match self {
+                    $(Self::$variant(s) => s.get_base(),)*
+                }
+            }
 
-    pub fn add_owner(&mut self, owner: Rid) {
-        *self.owners.entry(owner).or_insert(0) += 1;
-    }
+            fn get_mut_base(&mut self) -> &mut RapierShapeBase {
+                match self {
+                    $(Self::$variant(s) => s.get_mut_base(),)*
+                }
+            }
 
-    pub fn remove_owner(&mut self, owner: Rid) {
-        if let Some(count) = self.owners.get_mut(&owner) {
-            *count -= 1;
-            if *count == 0 {
-                self.owners.remove(&owner);
+            fn get_type(&self) -> ShapeType {
+                match self {
+                    $(Self::$variant(s) => s.get_type(),)*
+                }
+            }
+
+            fn allows_one_way_collision(&self) -> bool {
+                match self {
+                    $(Self::$variant(s) => s.allows_one_way_collision(),)*
+                }
+            }
+
+            fn create_rapier_shape(&mut self, physics_engine: &mut PhysicsEngine) -> ShapeHandle {
+                match self {
+                    $(Self::$variant(s) => s.create_rapier_shape(physics_engine),)*
+                }
+            }
+
+            fn set_data(&mut self, data: Variant, physics_engine: &mut PhysicsEngine) {
+                match self {
+                    $(Self::$variant(s) => s.set_data(data, physics_engine),)*
+                }
+            }
+
+            fn get_data(&self) -> Variant {
+                match self {
+                    $(Self::$variant(s) => s.get_data(),)*
+                }
+            }
+
+            fn get_handle(&self) -> ShapeHandle {
+                match self {
+                    $(Self::$variant(s) => s.get_handle(),)*
+                }
             }
         }
-    }
-
-    pub fn get_owners(&self) -> &HashMap<Rid, i32> {
-        &self.owners
-    }
-
-    pub fn get_rid(&self) -> Rid {
-        self.rid
-    }
-
-    pub fn destroy_shape(&mut self, physics_engine: &mut PhysicsEngine) {
-        if self.handle != ShapeHandle::default() {
-            physics_engine.shape_destroy(self.handle);
-            self.handle = ShapeHandle::default();
-        }
-    }
+    };
 }
-impl Drop for RapierShapeBase {
-    fn drop(&mut self) {
-        if !self.owners.is_empty() {
-            godot_error!("RapierShapeBase leaked {} owners", self.owners.len());
-        }
-        if self.is_valid() {
-            godot_error!("RapierShapeBase leaked");
-        }
-    }
-}
+#[cfg(feature = "dim3")]
+impl_rapier_shape_trait!(
+    RapierShape,
+    RapierCapsuleShape,
+    RapierCircleShape,
+    RapierConcavePolygonShape,
+    RapierConvexPolygonShape,
+    RapierCylinderShape3D,
+    RapierHeightMapShape3D,
+    RapierRectangleShape,
+    RapierSeparationRayShape,
+    RapierWorldBoundaryShape
+);
+#[cfg(feature = "dim2")]
+impl_rapier_shape_trait!(
+    RapierShape,
+    RapierCapsuleShape,
+    RapierCircleShape,
+    RapierConcavePolygonShape,
+    RapierConvexPolygonShape,
+    RapierSegmentShape2D,
+    RapierRectangleShape,
+    RapierSeparationRayShape,
+    RapierWorldBoundaryShape
+);
